+++ Craftsman/Builders/ExtensionBuilders/MassTransitExtensionsBuilder.cs
@@ -30,6 +30,7 @@ public class MassTransitExtensionsBuilder
 using {utilsClassPath.ClassNamespace};
 using {envServiceClassPath.ClassNamespace};
 using {messagesClassPath.ClassNamespace};
+using Configurations;
 using MassTransit;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
@@ -39,8 +40,10 @@ using System.Reflection;
 
 public static class MassTransitServiceExtension
 {{
-    public static void AddMassTransitServices(this IServiceCollection services, IWebHostEnvironment env)
+    public static void AddMassTransitServices(this IServiceCollection services, IWebHostEnvironment env, IConfiguration configuration)
     {{
+        var rmqOptions = configuration.GetRabbitMqOptions();
+
         if (!env.IsEnvironment(Consts.Testing.IntegrationTestingEnvName) 
             && !env.IsEnvironment(Consts.Testing.FunctionalTestingEnvName))
         {{
@@ -49,13 +52,13 @@ public static class MassTransitServiceExtension
                 mt.AddConsumers(Assembly.GetExecutingAssembly());
                 mt.UsingRabbitMq((context, cfg) =>
                 {{
-                    cfg.Host(EnvironmentService.RmqHost, 
-                        ushort.Parse(EnvironmentService.RmqPort), 
-                        EnvironmentService.RmqVirtualHost, 
+                    cfg.Host(rmqOptions.Host, 
+                        ushort.Parse(rmqOptions.Port), 
+                        rmqOptions.VirtualHost, 
                         h =>
                         {{
-                            h.Username(EnvironmentService.RmqUsername);
-                            h.Password(EnvironmentService.RmqPassword);
+                            h.Username(rmqOptions.Username);
+                            h.Password(rmqOptions.Password);
                         }});
 
                     // Producers -- Do Not Delete This Comment