+++ Craftsman/Services/FileNames.cs
@@ -46,7 +46,12 @@ public static class FileNames
         return $"{messageName}";
     }
 
+    public static string ConnectionStringOptionKey(string projectBaseName) => $"{projectBaseName}Key";
     public static string TestingServiceScope() => "TestingServiceScope";
+    public static string AuthOptions() => "AuthOptions";
+    public static string RabbitMqOptions() => "RabbitMqOptions";
+    public static string ConnectionStringOptions() => "ConnectionStringOptions";
+    public static string RootConfigurationExtensions() => "RootConfigurationExtensions";
     public static string FakeBuilderName(string entityName) => $"Fake{entityName}Builder";    
     public static string MessageInterfaceName(string messageName) => $"I{messageName}";
     public static string EntityCreatedDomainMessage(string entityName) => $"{entityName}Created";    
@@ -63,7 +68,7 @@ public static class FileNames
     public static string GetSeederName(Entity entity) => $"{entity.Name}Seeder";
     public static string GetInfraRegistrationName() => "InfrastructureServiceExtension";
     public static string GetSwaggerServiceExtensionName() => "SwaggerServiceExtension";
-    public static string GetAppSettingsName(bool asJson = true) => asJson ? $"appsettings.json" : $"appsettings";
+    public static string GetAppSettingsName(bool isDev = false) => isDev ? $"appsettings.Development.json" : $"appsettings.json";
     public static string BffApiKeysFilename(string entityName) => $"{entityName.LowercaseFirstLetter()}.keys";
     public static string BffEntityListRouteComponentName(string entityName) => $"{entityName.UppercaseFirstLetter()}List";
     public static string BffApiKeysExport(string entityName) => $"{entityName.UppercaseFirstLetter()}Keys";